+++ src/main/java/com/jomo/eorganism/metamodel/web/entity/DomainEntity.java
@@ -18,9 +18,9 @@ import java.util.Date;
 public class DomainEntity extends BaseEntity {
 
     private String uuid;
-    private Long metadataId;
-    private Long lastUpdatedUserId;
-    private Long lastUpdatedApplicationId;
+    private int metadataId;
+    private int lastUpdatedUserId;
+    private int lastUpdatedApplicationId;
 
     private String name;
     private String type;
@@ -120,27 +120,27 @@ public class DomainEntity extends BaseEntity {
         this.lastUpdatedApplicationName = lastUpdatedApplicationName;
     }
 
-    public Long getMetadataId() {
+    public int getMetadataId() {
         return metadataId;
     }
 
-    public void setMetadataId(Long metadataId) {
+    public void setMetadataId(int metadataId) {
         this.metadataId = metadataId;
     }
 
-    public Long getLastUpdatedUserId() {
+    public int getLastUpdatedUserId() {
         return lastUpdatedUserId;
     }
 
-    public void setLastUpdatedUserId(Long lastUpdatedUserId) {
+    public void setLastUpdatedUserId(int lastUpdatedUserId) {
         this.lastUpdatedUserId = lastUpdatedUserId;
     }
 
-    public Long getLastUpdatedApplicationId() {
+    public int getLastUpdatedApplicationId() {
         return lastUpdatedApplicationId;
     }
 
-    public void setLastUpdatedApplicationId(Long lastUpdatedApplicationId) {
+    public void setLastUpdatedApplicationId(int lastUpdatedApplicationId) {
         this.lastUpdatedApplicationId = lastUpdatedApplicationId;
     }
