+++ src/main/java/com/jomo/eorganism/metamodel/web/entity/EnvironmentEntity.java
@@ -15,10 +15,10 @@ import com.jomo.eorganism.metamodel.web.util.MetamodelUtil;
 public class EnvironmentEntity extends BaseEntity {
 
     private String uuid;
-    private Long releaseId;
-    private Long metadataId;
-    private Long lastUpdatedUserId;
-    private Long lastUpdatedApplicationId;
+    private int releaseId;
+    private int metadataId;
+    private int lastUpdatedUserId;
+    private int lastUpdatedApplicationId;
 
     private String name;
     private String type;
@@ -60,11 +60,11 @@ public class EnvironmentEntity extends BaseEntity {
         this.endDate = endDate;
     }
 
-    public Long getReleaseId() {
+    public int getReleaseId() {
         return releaseId;
     }
 
-    public void setReleaseId(Long releaseId) {
+    public void setReleaseId(int releaseId) {
         this.releaseId = releaseId;
     }
 
@@ -140,27 +140,27 @@ public class EnvironmentEntity extends BaseEntity {
         this.lastUpdatedApplicationName = lastUpdatedApplicationName;
     }
 
-    public Long getMetadataId() {
+    public int getMetadataId() {
         return metadataId;
     }
 
-    public void setMetadataId(Long metadataId) {
+    public void setMetadataId(int metadataId) {
         this.metadataId = metadataId;
     }
 
-    public Long getLastUpdatedUserId() {
+    public int getLastUpdatedUserId() {
         return lastUpdatedUserId;
     }
 
-    public void setLastUpdatedUserId(Long lastUpdatedUserId) {
+    public void setLastUpdatedUserId(int lastUpdatedUserId) {
         this.lastUpdatedUserId = lastUpdatedUserId;
     }
 
-    public Long getLastUpdatedApplicationId() {
+    public int getLastUpdatedApplicationId() {
         return lastUpdatedApplicationId;
     }
 
-    public void setLastUpdatedApplicationId(Long lastUpdatedApplicationId) {
+    public void setLastUpdatedApplicationId(int lastUpdatedApplicationId) {
         this.lastUpdatedApplicationId = lastUpdatedApplicationId;
     }
