+++ src/main/java/com/jomo/eorganism/metamodel/web/repository/MetadataRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface MetadataRepository extends JpaRepository<MetadataEntity, Long> {
+public interface MetadataRepository extends JpaRepository<MetadataEntity, Integer> {
 }
