+++ src/main/java/com/jomo/eorganism/metamodel/web/service/DomainService.java
@@ -8,7 +8,7 @@ public interface DomainService {
     Iterable<DomainEntity> listDomains();
 
     @TransactionalReadOnly
-    DomainEntity findDomain(long id);
+    DomainEntity findDomain(int id);
 }
