+++ src/main/java/com/jomo/eorganism/metamodel/web/service/MetadataService.java
@@ -8,7 +8,7 @@ public interface MetadataService {
     Iterable<MetadataEntity> listMetadatas();
 
     @TransactionalReadOnly
-    MetadataEntity findMetadata(long id);
+    MetadataEntity findMetadata(int id);
 }
